Move a bunch more allocation tests out of runtests.c
