+++ expat/tests/handlers.h
@@ -382,6 +382,12 @@ extern int XMLCALL external_entity_dbl_handler_2(XML_Parser parser,
                                                  const XML_Char *systemId,
                                                  const XML_Char *publicId);
 
+extern int XMLCALL external_entity_alloc_set_encoding(XML_Parser parser,
+                                                      const XML_Char *context,
+                                                      const XML_Char *base,
+                                                      const XML_Char *systemId,
+                                                      const XML_Char *publicId);
+
 /* NotStandalone handlers */
 
 extern int XMLCALL reject_not_standalone_handler(void *userData);
